+++ sekoia_automation/trigger.py
@@ -3,16 +3,16 @@ import signal
 from abc import abstractmethod
 from collections.abc import Generator
 from contextlib import contextmanager
-from datetime import datetime, timedelta
+from datetime import UTC, datetime, timedelta
 from http.server import BaseHTTPRequestHandler, HTTPServer
 from pathlib import Path
 from threading import Event, Thread
 from typing import Any, ClassVar
 from uuid import uuid4
 
+import botocore.exceptions
 import requests
 import sentry_sdk
-from botocore.exceptions import ClientError, ConnectionError, HTTPClientError
 from cachetools import TLRUCache
 from pydantic.v1 import BaseModel
 from requests import HTTPError
@@ -69,9 +69,9 @@ class Trigger(ModuleItem):
         super().__init__(module, data_path)
         self._configuration: dict | BaseModel | None = None
         self._error_count = 0
-        self._last_events_time = datetime.utcnow()
-        self._last_heartbeat = datetime.utcnow()
-        self._startup_time = datetime.utcnow()
+        self._last_events_time = datetime.now(UTC).replace(tzinfo=None)
+        self._last_heartbeat = datetime.now(UTC).replace(tzinfo=None)
+        self._startup_time = datetime.now(UTC).replace(tzinfo=None)
         sentry_sdk.set_tag("item_type", "trigger")
         self._secrets: dict[str, Any] = {}
         self._stop_event = Event()
@@ -167,11 +167,14 @@ class Trigger(ModuleItem):
         except (TriggerConfigurationError, ModuleConfigurationError) as e:
             self.log_exception(e)
             self.log("Configuration error", "critical")
-        except (ConnectionError, HTTPClientError) as ex:
+        except (
+            botocore.exceptions.ConnectionError,
+            botocore.exceptions.HTTPClientError,
+        ) as ex:
             # Error while communicating with the S3 storage
             # Don't increment the error count because this is an internal issue
             self.log_exception(ex)
-        except ClientError as ex:
+        except botocore.exceptions.ClientError as ex:
             self._handle_s3_exception(ex)
         except SendEventError as ex:
             self._handle_send_event_exception(ex)
@@ -205,7 +208,7 @@ class Trigger(ModuleItem):
         """
         Mark the trigger as alive.
         """
-        self._last_heartbeat = datetime.utcnow()
+        self._last_heartbeat = datetime.now(UTC).replace(tzinfo=None)
 
     def _rm_tree(self, path: Path) -> None:
         """
@@ -278,7 +281,7 @@ class Trigger(ModuleItem):
         """
         # Reset the consecutive error count
         self._error_count = 0
-        self._last_events_time = datetime.utcnow()
+        self._last_events_time = datetime.now(UTC).replace(tzinfo=None)
 
         # Add request_id to the data to be able to track the request
         data = {"name": event_name, "event": event, "request_id": str(uuid4())}
@@ -340,7 +343,7 @@ class Trigger(ModuleItem):
 
         self._logs.append(
             {
-                "date": datetime.utcnow().isoformat(),
+                "date": datetime.now(UTC).replace(tzinfo=None).isoformat(),
                 "level": level,
                 "message": message,
             }
@@ -406,6 +409,7 @@ class Trigger(ModuleItem):
     def stop_monitoring(self):
         if self._liveness_server:
             self._liveness_server.shutdown()
+            self._liveness_server.server_close()
             self._liveness_server = None
 
         # Stop the exporter
@@ -428,7 +432,7 @@ class Trigger(ModuleItem):
         return True
 
     def _events_alive(self) -> bool:
-        delta = datetime.utcnow() - self._last_events_time
+        delta = datetime.now(UTC).replace(tzinfo=None) - self._last_events_time
         if self.seconds_without_events <= 0 or delta < timedelta(
             seconds=self.seconds_without_events
         ):
@@ -443,7 +447,7 @@ class Trigger(ModuleItem):
         return False
 
     def _heartbeat_alive(self) -> bool:
-        heartbeat_delta = datetime.utcnow() - self._last_heartbeat
+        heartbeat_delta = datetime.now(UTC).replace(tzinfo=None) - self._last_heartbeat
         if self.last_heartbeat_threshold <= 0 or heartbeat_delta < timedelta(
             seconds=self.last_heartbeat_threshold
         ):
@@ -499,17 +503,18 @@ class Trigger(ModuleItem):
             return False
 
         delta_since_last_event = (
-            datetime.utcnow() - self._last_events_time
+            datetime.now(UTC).replace(tzinfo=None) - self._last_events_time
         ).total_seconds()
         delta_since_startup = min(
-            datetime.utcnow() - self._startup_time, timedelta(days=5)
+            datetime.now(UTC).replace(tzinfo=None) - self._startup_time,
+            timedelta(days=5),
         ).total_seconds()
         if delta_since_startup < 1800:
             # Graceful 30 minutes period at startup
             return False
         return delta_since_startup / 5 <= delta_since_last_event
 
-    def _handle_s3_exception(self, ex: ClientError):
+    def _handle_s3_exception(self, ex: botocore.exceptions.ClientError):
         """
         Handle errors coming from the S3 storage
         """
@@ -545,7 +550,7 @@ class Trigger(ModuleItem):
 class LivenessHandler(BaseHTTPRequestHandler):
     trigger: Trigger
 
-    def do_GET(self):  # noqa: N802
+    def do_GET(self):
         if self.path == "/health":
             self.send(
                 200 if self.trigger.is_alive() else 500, self.trigger.liveness_context()