+++ tests/test_trigger.py
@@ -1,15 +1,14 @@
 import datetime
 import time
-from datetime import timedelta
+from datetime import UTC, timedelta
 from pathlib import Path
 from typing import ClassVar
 from unittest.mock import PropertyMock, mock_open, patch
 
-# third parties
+import botocore.exceptions
 import pytest
 import requests
 import requests_mock
-from botocore.exceptions import ClientError, ConnectionError
 from pydantic.v1 import BaseModel
 from tenacity import wait_none
 
@@ -448,8 +447,12 @@ def test_too_many_errors_critical_log(_, mocked_trigger_logs):
 
     trigger = TestTrigger()
     trigger._error_count = 4
-    trigger._startup_time = datetime.datetime.utcnow() - timedelta(hours=1)
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=5)
+    trigger._startup_time = datetime.datetime.now(UTC).replace(tzinfo=None) - timedelta(
+        hours=1
+    )
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=5)
     trigger._STOP_EVENT_WAIT = 0.001
     with (
         pytest.raises(SystemExit),
@@ -530,9 +533,9 @@ def test_trigger_liveness(monitored_trigger):
 
 def test_trigger_liveness_error(monitored_trigger, mocked_trigger_logs):
     monitored_trigger.seconds_without_events = 1
-    monitored_trigger._last_events_time = (
-        datetime.datetime.utcnow() - datetime.timedelta(seconds=60)
-    )
+    monitored_trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - datetime.timedelta(seconds=60)
     mocked_trigger_logs.register_uri(
         "GET", "http://127.0.0.1:8000/health", real_http=True
     )
@@ -546,9 +549,9 @@ def test_trigger_liveness_error(monitored_trigger, mocked_trigger_logs):
 
 def test_trigger_liveness_heartbeat_error(monitored_trigger, mocked_trigger_logs):
     monitored_trigger.last_heartbeat_threshold = 1
-    monitored_trigger._last_heartbeat = datetime.datetime.utcnow() - datetime.timedelta(
-        seconds=60
-    )
+    monitored_trigger._last_heartbeat = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - datetime.timedelta(seconds=60)
     mocked_trigger_logs.register_uri(
         "GET", "http://127.0.0.1:8000/health", real_http=True
     )
@@ -567,7 +570,7 @@ def test_trigger_liveness_not_found(monitored_trigger):
 
 def test_trigger_s3_connection_error(mocked_trigger_logs):
     trigger = ErrorTrigger()
-    trigger.ex = ConnectionError(error="Err")
+    trigger.ex = botocore.exceptions.ConnectionError(error="Err")
 
     with patch("sentry_sdk.capture_exception") as sentry_patch:
         trigger._execute_once()
@@ -578,7 +581,7 @@ def test_trigger_s3_connection_error(mocked_trigger_logs):
 
 def test_trigger_s3_server_error_int(mocked_trigger_logs):
     trigger = ErrorTrigger()
-    trigger.ex = ClientError({"Error": {"Code": 500}}, "foo")
+    trigger.ex = botocore.exceptions.ClientError({"Error": {"Code": 500}}, "foo")
     with patch("sentry_sdk.capture_exception") as sentry_patch:
         trigger._execute_once()
         sentry_patch.assert_called()
@@ -588,7 +591,9 @@ def test_trigger_s3_server_error_int(mocked_trigger_logs):
 
 def test_trigger_s3_server_error_str(mocked_trigger_logs):
     trigger = ErrorTrigger()
-    trigger.ex = ClientError({"Error": {"Code": "ServiceUnavailable"}}, "foo")
+    trigger.ex = botocore.exceptions.ClientError(
+        {"Error": {"Code": "ServiceUnavailable"}}, "foo"
+    )
     with patch("sentry_sdk.capture_exception") as sentry_patch:
         trigger._execute_once()
         sentry_patch.assert_called()
@@ -598,7 +603,7 @@ def test_trigger_s3_server_error_str(mocked_trigger_logs):
 
 def test_trigger_s3_client_error_int(mocked_trigger_logs):
     trigger = ErrorTrigger()
-    trigger.ex = ClientError({"Error": {"Code": 400}}, "foo")
+    trigger.ex = botocore.exceptions.ClientError({"Error": {"Code": 400}}, "foo")
     with patch("sentry_sdk.capture_exception") as sentry_patch:
         trigger._execute_once()
         sentry_patch.assert_called()
@@ -609,7 +614,9 @@ def test_trigger_s3_client_error_int(mocked_trigger_logs):
 
 def test_trigger_s3_client_error_str(mocked_trigger_logs):
     trigger = ErrorTrigger()
-    trigger.ex = ClientError({"Error": {"Code": "NoSuchBucket"}}, "foo")
+    trigger.ex = botocore.exceptions.ClientError(
+        {"Error": {"Code": "NoSuchBucket"}}, "foo"
+    )
     with patch("sentry_sdk.capture_exception") as sentry_patch:
         trigger._execute_once()
         sentry_patch.assert_called()
@@ -642,8 +649,12 @@ def test_is_error_critical_errors():
     assert trigger._is_error_critical() is False
     trigger._error_count = 5
     assert trigger._is_error_critical() is False
-    trigger._startup_time = datetime.datetime.utcnow() - timedelta(hours=1)
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=9)
+    trigger._startup_time = datetime.datetime.now(UTC).replace(tzinfo=None) - timedelta(
+        hours=1
+    )
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=9)
     assert trigger._is_error_critical() is True
 
 
@@ -655,24 +666,38 @@ def test_is_error_critical_time_since_last_event():
     trigger._startup_time = datetime.datetime(year=2021, month=1, day=1)
     assert trigger._is_error_critical() is False
     # Time without events is capped to 24 hours
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=23)
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=23)
     assert trigger._is_error_critical() is False
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=24)
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=24)
     assert trigger._is_error_critical() is True
 
     # Trigger that just started and already has 5 errors without sending any event
     trigger = DummyTrigger()
     trigger._error_count = 5
     assert trigger._is_error_critical() is False
-    trigger._startup_time = datetime.datetime.utcnow() - timedelta(hours=1)
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=5)
+    trigger._startup_time = datetime.datetime.now(UTC).replace(tzinfo=None) - timedelta(
+        hours=1
+    )
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=5)
     assert trigger._is_error_critical() is True
 
     # Trigger that has been running for 1 day should exit after 5 hours of errors
-    trigger._last_events_time = datetime.datetime.utcnow()
-    trigger._startup_time = datetime.datetime.utcnow() - timedelta(days=1)
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(tzinfo=None)
+    trigger._startup_time = datetime.datetime.now(UTC).replace(tzinfo=None) - timedelta(
+        days=1
+    )
     assert trigger._is_error_critical() is False
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=1)
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=1)
     assert trigger._is_error_critical() is False
-    trigger._last_events_time = datetime.datetime.utcnow() - timedelta(hours=5)
+    trigger._last_events_time = datetime.datetime.now(UTC).replace(
+        tzinfo=None
+    ) - timedelta(hours=5)
     assert trigger._is_error_critical() is True